requirements.txt
fdb==2.0.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
PyYAML==6.0.1
